+++ src/app/app.component.ts
@@ -19,7 +19,6 @@ export class AppComponent {
     decimal: ',',
     allowNegative: true,
     nullable: true,
-    inputMode: ResolveCurrencyMaskInputMode.Natural,
     max: 250_000_000,
     inputMode: ResolveCurrencyMaskInputMode.Financial,
   });
@@ -127,6 +126,7 @@ bootstrapApplication(AppComponent, {
 
   constructor(private readonly formBuilder: FormBuilder) {
     this.form.controls.inputMode.valueChanges.subscribe(inputMode => {
+      console.log(inputMode)
       this.resolveCurrencyMaskOptions.update(options => ({
         ...options,
         inputMode,